グループ追加処理見直し

+++ server/app/Http/Controllers/Api/InvitationController.php
@@ -3,15 +3,14 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use App\Models\Group;
+use App\Models\GroupUserMapping;
 use App\Models\InvitationToken;
-use App\Models\ShoppingCategory;
-use App\Models\ShoppingItem;
 use Illuminate\Http\Request;
 use Exception;
 use Carbon\Carbon;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Log;
 
 class InvitationController extends Controller
 {
@@ -65,7 +64,7 @@ class InvitationController extends Controller
         });
 
         return response()->json([
-            'token' => $invitationToken->token,
+            'token' => $token,
             'expires_at' => $invitationToken->expires_at,
             'inviter' => [
                 'id' => $invitationToken->inviter->id,
@@ -127,7 +126,7 @@ class InvitationController extends Controller
         }
 
         // 招待された人がすでに同じグループにいるかチェック
-        if ($currentGroup->id === $inviter->group_id) {
+        if ($currentGroup->id === $inviter->group->id) {
             return response()->json(['message' => 'すでにグループに参加しています。'], 403);
         }
 
@@ -150,14 +149,28 @@ class InvitationController extends Controller
 
         // 招待された人を同じグループに追加
         $group = $inviter->group;
-        $user->groupUser->group_id = $group->id;
-        $user->groupUser->save();
+        Log::info('invitation_controller.join', ['group' => $group, 'user' => $user, 'inviter' => $inviter]);
+
+        // 既存のグループユーザーマッピングを削除
+        $user->groupUser()->delete();
+
+        // 新しいグループユーザーマッピングを作成
+        GroupUserMapping::create([
+            'user_id' => $user->id,
+            'group_id' => $group->id
+        ]);
 
         $group->group_size = $group->getGroupSize();
         $group->save();
 
-        // // 招待された人のグループを削除
-        $currentGroup->delete();
+        // 現在のグループのサイズを更新
+        $currentGroup->group_size = $currentGroup->getGroupSize();
+        $currentGroup->save();
+
+        // 現在のグループに所属するユーザーが0人になった場合のみグループを削除
+        if ($currentGroup->group_size === 0) {
+            $currentGroup->delete();
+        }
 
         return response()->json(['message' => 'グループに参加しました。'], 200);
     }
